Most likely 817fdad uncovered the fact that our choice of primary replica was resonating with tablet allocation and we were ending up picking the same replica as primary within a scope instead of rotating primaryship among all replicas in the scope.
This created situations where for instance, restoring into a 9 nodes with primary_replica_only=true would put all data into 3 nodes, leaving the other 6 unused. The balancing of the dataset was performed by the subsequent repair step.

This PR fixes this by changing the formula for picking up the primary replica out of a set of eligible replicas from within the passed scope.
The PR also extends the testing scenarios in `test_backup.py` so we get to run restore for a set of topologies, for all combinations of scope, primary_replica_only and min_tablet_counts.
Most of the work was done by @bhalevy [here](https://github.com/scylladb/scylladb/compare/master...bhalevy:scylla:load-balance-primary-replica), this PR just splitted it and did touchups here and there.

Fixes #27281

Closes scylladb/scylladb#27397

* github.com:scylladb/scylladb:
  test: reduce dataset and number of test cases or debug builds
  test: bump repair timeout up, it's sometimes not enough in CI
  test: refactor test_refresh.py to match test_restore_with_streaming_scopes.
  test: extend test_restore_with_streaming_scopes
  test: Adjust test_restore_primary_replica_different_dc_scope_all
  test: Refactor restoring code in test_backup to match SM pattern
  test: add check_mutation_replicas calls after fresh creation of dataset
  test: extend create_dataset to accept consistency_level
  test: refactor check_mutation_replicas so it's more readable
  test: make create_dataset async and refactor so it's configurable
  test: use defaultdict in collect_mutations
  test: add log marks to facilitate reusing server for restore
  locator: tablets: Distribute data evenly among primary replicas during restore

+++ test/cluster/test_refresh.py
@@ -19,7 +19,7 @@ from collections import defaultdict
 from test.pylib.minio_server import MinioServer
 from test.pylib.manager_client import ManagerClient
 from test.cluster.object_store.conftest import format_tuples
-from test.cluster.object_store.test_backup import topo, create_cluster, take_snapshot, create_dataset, compute_scope, check_data_is_back, check_mutation_replicas
+from test.cluster.object_store.test_backup import topo, create_cluster, take_snapshot, create_dataset, check_data_is_back, do_load_sstables, mark_all_logs, check_mutation_replicas
 from test.cluster.util import wait_for_cql_and_get_hosts
 from test.pylib.rest_client import read_barrier
 from test.pylib.util import unique_name
@@ -64,7 +64,10 @@ async def test_refresh_with_streaming_scopes(manager: ManagerClient, topology_rf
 
     ks = 'ks'
     cf = 'cf'
-    _, keys, _ = create_dataset(manager, ks, cf, topology, logger)
+    _, keys, _ = await create_dataset(manager, ks, cf, topology, logger, num_keys=10000, min_tablet_count=512)
+
+    # validate replicas assertions hold on fresh dataset
+    await check_mutation_replicas(cql, manager, servers, keys, topology, logger, ks, cf, scope=None, primary_replica_only=False, expected_replicas = None)
 
     _, sstables = await take_snapshot(ks, servers, manager, logger)
 
@@ -86,29 +89,44 @@ async def test_refresh_with_streaming_scopes(manager: ManagerClient, topology_rf
             if item not in exclude_list:
                 shutil.copy2(src_path, dst_path)
 
-    logger.info(f'Clear data by truncating, make sure the tablets map stays intact')
-    cql.execute(f'TRUNCATE TABLE {ks}.{cf};')
-
-    logger.info(f'Move sstables to upload dir')
-    for i, s in enumerate(servers):
-        workdir = await manager.server_get_workdir(s.server_id)
-
-        cf_dir = os.listdir(f'{workdir}/data/{ks}')[0]
-        cf_dir = os.path.join(f'{workdir}/data/{ks}', cf_dir)
-
-        tmpbackup = os.path.join(workdir, f'../{tmpdir}')
-        shutil.copytree(tmpbackup, os.path.join(cf_dir, 'upload'), dirs_exist_ok=True)
-
     logger.info(f'Refresh')
-    async def do_refresh(s, toc_names, scope):
+    async def do_refresh(manager, logger, ks, cf, s, toc_names, scope, primary_replica_only, _prefix=None, _object_storage=None):
+        # Get the list of toc_names that this node needs to load and find all sstables
+        # that correspond to these toc_names, copy them to the upload directory and then
+        # call refresh
+        workdir = await manager.server_get_workdir(s.server_id)
+        cf_dir = os.listdir(f'{workdir}/data/{ks}')[0]
+        upload_dir = os.path.join(f'{workdir}/data/{ks}', cf_dir, 'upload')
+        os.makedirs(upload_dir, exist_ok=True)
+        tmpbackup = os.path.join(workdir, f'../{tmpdir}')
+        for toc in toc_names:
+            basename = toc.removesuffix('-TOC.txt')
+            for item in os.listdir(tmpbackup):
+                if item.startswith(basename):
+                    src_path = os.path.join(tmpbackup, item)
+                    dst_path = os.path.join(upload_dir, item)
+                    shutil.copy2(src_path, dst_path)
+
         logger.info(f'Refresh {s.ip_addr} with {toc_names}, scope={scope}')
-        await manager.api.load_new_sstables(s.ip_addr, ks, cf, scope=scope, load_and_stream=True)
+        await manager.api.load_new_sstables(s.ip_addr, ks, cf, scope=scope, primary_replica=primary_replica_only, load_and_stream=True)
 
-    scope,r_servers = compute_scope(topology, servers)
+    for scope in ['all', 'dc', 'rack', 'node']:
+        # We can support rack-aware restore with rack lists, if we restore the rack-list per dc as it was at backup time.
+        # Otherwise, with numeric replication_factor we'd pick arbitrary subset of the racks when the keyspace
+        # is initially created and an arbitrary subset or the rack at restore time.
+        if scope == 'rack' and topology.rf != topology.racks:
+            logger.info(f'Skipping scope={scope} test since rf={topology.rf} != racks={topology.racks} and it cannot be supported with numeric replication_factor')
+            continue
+        pros = [False] if scope == 'node' else [False, True]
+        for pro in pros:
+            logger.info(f'Clear data by truncating, make sure the tablets map stays intact')
+            cql.execute(f'TRUNCATE TABLE {ks}.{cf};')
 
-    await asyncio.gather(*(do_refresh(s, sstables, scope) for s in r_servers))
+            log_marks = await mark_all_logs(manager, servers)
 
-    await check_data_is_back(manager, logger, cql, ks, cf, keys, servers, topology, r_servers, host_ids, scope)
+            await do_load_sstables(ks, cf, servers, topology, sstables, scope, manager, logger, primary_replica_only=pro, load_fn=do_refresh)
+
+            await check_data_is_back(manager, logger, cql, ks, cf, keys, servers, topology, host_ids, scope, primary_replica_only=pro, log_marks=log_marks)
 
     shutil.rmtree(tmpbackup)
 
@@ -127,7 +145,7 @@ async def test_refresh_deletes_uploaded_sstables(manager: ManagerClient):
 
     ks = 'ks'
     cf = 'cf'
-    _, keys, _ = create_dataset(manager, ks, cf, topology, logger)
+    _, keys, _ = await create_dataset(manager, ks, cf, topology, logger)
 
     _, sstables = await take_snapshot(ks, servers, manager, logger)
